Moving bb files to a separate folder
