+++ pysixtrack/be_beambeam/pyfaddeeva.py
@@ -1,6 +1,4 @@
-# attaching faddeeva to np
 from scipy.special import wofz
-
 
 def wfun(z_re, z_im):
     w = wofz(z_re+1j*z_im)
