+++ pysixtrack/be_beambeam/slicing.py
